% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_abs_circle.R
\name{geom_abs_circle}
\alias{geom_abs_circle}
\title{A 'ggplot2' plot to draw a circle with absolute radius}
\usage{
geom_abs_circle(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  radius = grid::unit(50, "mm"),
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for \code{ggplot2::geom_text()}.}

\item{radius}{Radius of the circle. A \code{grid::unit()} object. Defaults to 50
mm.}
}
\description{
A 'ggplot2' plot to draw a circle with absolute radius
}
\section{Aesthetics}{

\itemize{
\item x (required)
\item y (required)
\item alpha
\item colour
\item fill
\item linetype
\item linewidth
}
}

