% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circumscribe.R
\name{geom_circumscribe}
\alias{geom_circumscribe}
\title{A 'ggplot2' geom to fit text inside a circle}
\usage{
geom_circumscribe(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  padding = grid::unit(4, "mm"),
  grow = FALSE,
  reflow = FALSE,
  ...
)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{Standard
geom arguments as for \code{ggplot2::geom_text()}.}

\item{padding}{Padding between the text and the circle. A \code{grid::unit()}
object. Defaults to \code{grid::unit(4, "mm")}.}

\item{grow}{If \code{TRUE}, text will be made as large as able to fit in the
circle.}

\item{reflow}{If \code{TRUE}, text will be reflowed (wrapped) to fill the circle
as tightly as possible.}
}
\description{
\code{geom_circumscribe()} shrinks, grows, and wraps text to fit inside a circle.
}
\details{
Except where noted, \code{geom_circumscribe()} behaves like
\code{ggplot2::geom_text()}.

If the text is too big to fit in the circle, it will shrink to fit the
circle. If \code{grow = TRUE} is set, the text will be made as large as possible
whether that means shrinking or growing it.

If \code{reflow = TRUE} is set, the text will be reflowed (wrapped) to fill the
circle as tightly as possible. If the text is still too big for the circle
after reflowing, it will be shrunk as usual. Existing line breaks are
respected when reflowing.

TODO need to figure out a principled mapping of data to radius, or
alternatively another aesthetic that could reasonably be mapped to radius.
}
\section{Aesthetics}{

\itemize{
\item label (required)
\item x (required)
\item y (required)
\item radius (required)
\item alpha
\item angle
\item colour
\item family
\item fontface
\item lineheight
\item size
}
}

